@@ -13,28 +13,30 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^5.0.0",
-    "@angular/common": "^5.0.0",
-    "@angular/compiler": "^5.0.0",
-    "@angular/core": "^5.0.0",
-    "@angular/forms": "^5.0.0",
-    "@angular/http": "^5.0.0",
-    "@angular/platform-browser": "^5.0.0",
-    "@angular/platform-browser-dynamic": "^5.0.0",
-    "bootstrap": "^4.0.0-beta.3",
+    "@angular/animations": "^6.0.0",
+    "@angular/common": "^6.0.0",
+    "@angular/compiler": "^6.0.0",
+    "@angular/core": "^6.0.0",
+    "@angular/forms": "^6.0.0",
+    "@angular/http": "^6.0.0",
+    "@angular/platform-browser": "^6.0.0",
+    "@angular/platform-browser-dynamic": "^6.0.0",
+    "bootstrap": "^4.1.1",
     "core-js": "^2.4.1",
     "popper.js": "^1.11.0",
+    "rxjs": "^6.1.0",
     "tether": "^1.4.4",
     "ts-helpers": "^1.1.1",
     "zone.js": "^0.8.26"
   },
   "devDependencies": {
-    "@angular/cli": "^1.7.4",
-    "@angular/compiler-cli": "^5.0.0",
-    "@angular/language-service": "^5.0.0",
-    "@types/jasmine": "^2.2.30",
+    "@angular-devkit/build-angular": "~0.6.0",
+    "@angular/cli": "^6.0.0",
+    "@angular/compiler-cli": "^6.0.0",
+    "@angular/language-service": "^6.0.0",
+    "@types/jasmine": "^2.8.7",
     "@types/jasminewd2": "~2.0.2",
-    "@types/node": "^9.6.7",
+    "@types/node": "^10.0.3",
     "codelyzer": "~4.3.0",
     "jasmine-core": "3.1.0",
     "jasmine-spec-reporter": "4.2.1",
@@ -43,13 +45,13 @@
     "karma-chrome-launcher": "^2.0.0",
     "karma-cli": "^1.0.1",
     "karma-coverage-istanbul-reporter": "^1.2.1",
-    "karma-jasmine": "^1.0.2",
+    "karma-jasmine": "^1.1.2",
     "karma-jasmine-html-reporter": "^1.0.0",
     "karma-phantomjs-launcher": "^1.0.4",
     "karma-remap-istanbul": "^0.6.0",
     "protractor": "^5.3.1",
-    "ts-node": "^6.0.1",
-    "tslint": "5.9.1",
-    "typescript": "^2.6.2"
+    "ts-node": "^6.0.2",
+    "tslint": "^5.10.0",
+    "typescript": "^2.7.2"
   }
 }
